+++ src/dk/jonaslindstrom/mosef/modules/filter/filters/DiscreteFilterFactory.java
@@ -2,6 +2,6 @@ package dk.jonaslindstrom.mosef.modules.filter.filters;
 
 public interface DiscreteFilterFactory {
 
-  public float[] getSamples();
+  public double[] getSamples();
 
 }
