+++ src/dk/jonaslindstrom/mosef/modules/oscillator/waves/SquareWave.java
@@ -9,7 +9,7 @@ public class SquareWave extends SimpleWave {
   }
 
   @Override
-  public float getSample(float t) {
+  public double getSample(double t) {
     if (t < 0.5f) {
       return .99f;
     } else {
